Add files via upload
This will extract the lyrics from the selected notes, purge any "br" or "hh" etc., and put the result on the clipboard. If there ae uppercase letters in the text, it will line break on them. (This presents a problem with "I" and proper names. You'll have to clean those up manually.)

It will put in an extra line break if there is a blank measure between lyrics.

No idea if or how this will work in languages other than English.