Add files via upload
This script will run through your notes and clean up any overlaps. It will give you a choice to keep the length of the first note, the second note, or split the difference. Also, the script will remember settings between runs. To make it automatic, set the parameters as desired and check the box "Save as defaults and don't prompt again." This way you can assign it to a key you won't have to bother with the dialog box.

If you do that and you change your mind, you'll need to find and delete the file "fixoverlaps.cfg." On Windows, it's usually at "C:\Program Files\Synthesizer V Studio Pro\fixoverlaps.cfg," and on Mac, in your Home folder.